requirements.txt
anthropic>=0.40.0
openai>=1.0.0
pyyaml>=6.0
python-dotenv>=1.0.0
pygetwindow>=0.0.9
pynput>=1.7.6
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
websockets>=12.0
